refactor(cross-chain): flatten protocol subdirectories to flat files
Flatten single-file subdirectories in protocols/ to match simpler
flat file structure:
- across/constants/across.ts → across/constants.ts
- across/schemas/across.ts → across/schemas.ts
- across/schemas/acrossDiscovery.ts → across/discoverySchemas.ts
- across/types/across.ts → across/types.ts
- across/utils/acrossCalldataDecoder.ts → across/utils.ts
- across/validators/acrossPayloadValidator.ts → across/validator.ts
- oif/constants/openIntentFramework.ts → oif/constants.ts
- oif/schemas/oif.ts → oif/schemas.ts

Remove unnecessary barrel index.ts files for each subdirectory.
