+++ packages/cross-chain/src/core/interfaces/OifProvider.interface.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
 
-import { OifProviderConfigSchema } from "../../protocols/oif/schemas/oif.js";
+import { OifProviderConfigSchema } from "../../protocols/oif/schemas.js";
 
 export type OifProviderConfig = z.infer<typeof OifProviderConfigSchema>;
