--- packages/cross-chain/src/protocols/across/constants/index.ts
@@ -1,1 +1,0 @@
-export * from "./across.js";
