+++ packages/cross-chain/src/protocols/across/index.ts
@@ -1,6 +1,7 @@
 export * from "./provider.js";
-export * from "./validators/index.js";
-export * from "./utils/index.js";
-export * from "./schemas/index.js";
-export * from "./constants/index.js";
-export * from "./types/index.js";
+export * from "./validator.js";
+export * from "./utils.js";
+export * from "./schemas.js";
+export * from "./discoverySchemas.js";
+export * from "./constants.js";
+export * from "./types.js";
