--- packages/cross-chain/src/protocols/across/schemas/index.ts
@@ -1,2 +1,0 @@
-export * from "./across.js";
-export * from "./acrossDiscovery.js";
