--- packages/cross-chain/src/protocols/across/types/index.ts
@@ -1,1 +1,0 @@
-export * from "./across.js";
