--- packages/cross-chain/src/protocols/across/utils/index.ts
@@ -1,1 +1,0 @@
-export * from "./acrossCalldataDecoder.js";
