--- packages/cross-chain/src/protocols/across/validators/index.ts
@@ -1,1 +1,0 @@
-export * from "./acrossPayloadValidator.js";
