--- packages/cross-chain/src/protocols/oif/constants/index.ts
@@ -1,1 +1,0 @@
-export * from "./openIntentFramework.js";
