--- packages/cross-chain/src/protocols/oif/schemas/index.ts
@@ -1,1 +1,0 @@
-export * from "./oif.js";
